chore(executor): extract prefetch.rs (#116 step 6)
Pulls the tri-dialect prefetch_related helpers out of mod.rs:

- fetch_with_prefetch_pool — bi-dialect counterpart of the PG
  fetch_with_prefetch.
- fetch_with_prefetch_filtered_pool — Django's Prefetch(queryset=...)
  shape (issue #298 / T2.1).

The PG-only `_on` variants (annotate_count_children,
annotate_count_children_on, fetch_with_prefetch) plus the HasPkValue
trait stay in mod.rs — they're tightly coupled to PG's PgRow / PgPool
types and the schema-mode tenancy connection path. `extract_pk_value`
is now pub(super) so prefetch.rs can reach it.

mod.rs: 3640 → 3412 LOC (-228). prefetch.rs: 235 LOC.

1482 lib tests pass on the sqlite,tenancy litmus.
